backend/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
attrs==23.1.0
blinker==1.7.0
certifi==2024.2.2
click==8.1.7
distro==1.8.0
dnspython==2.6.1
email_validator==2.1.1
fastapi==0.110.0
Flask==3.0.0
h11==0.14.0
httpcore==1.0.1
httptools==0.6.1
httpx==0.25.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.26.4
openai==1.1.1
opencv-python==4.9.0.80
orjson==3.9.15
outcome==1.3.0.post0
pillow==10.2.0
pydantic==2.4.2
pydantic-extra-types==2.6.0
pydantic-settings==2.2.1
pydantic_core==2.10.1
PySocks==1.7.1
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
selenium==4.15.0
sniffio==1.3.0
sortedcontainers==2.4.0
starlette==0.36.3
tqdm==4.66.1
trio==0.22.2
trio-websocket==0.11.1
typing_extensions==4.8.0
ujson==5.9.0
urllib3==2.0.7
uvicorn==0.28.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.1
wsproto==1.2.0
